Projeto/requirements.txt
# padrão necessário 

asgiref==3.8.1
Django>=5.0.4
sqlparse==0.4.4
tzdata==2024.1

python-dotenv
psycopg2-binary
gunicorn
whitenoise

# extras para o projeto
Pillow
setuptools
django-notifications